+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## [1.3]
 
 ### - 2024-09-03
-- ArrayHelper class v1.3.2 업데이트 map,reduce 메소드 기능 추가
+- ArrayHelper class v1.3.2 업데이트 map,reduce,__set 메소드 기능 추가, extractValues -> pluck 으로 메소드명 변경
 
 ### - 2024-09-02
 - BaseAdapterInterface 추가